refactor: rename web3-react to web3Modal
