+++ libs/wallet/src/index.ts
@@ -4,18 +4,18 @@ export * from './api/types'
 
 // Hooks
 export * from './api/hooks'
-export * from './web3-react/hooks/useWalletMetadata'
-export * from './web3-react/hooks/useIsWalletConnect'
-export * from './web3-react/hooks/useIsMetaMask'
-export * from './web3-react/hooks/useSafeAppsSdk'
-export * from './web3-react/hooks/useIsSmartContractWallet'
+export * from './web3modal/hooks/useWalletMetadata'
+export * from './web3modal/hooks/useIsWalletConnect'
+export * from './web3modal/hooks/useIsMetaMask'
+export * from './web3modal/hooks/useSafeAppsSdk'
+export * from './web3modal/hooks/useIsSmartContractWallet'
 
 // Updater
-export * from './web3-react/updater'
+export * from './web3modal/updater'
 
 // Components
 export * from './api/container/Identicon'
-export * from './web3-react/pure/AccountIndexSelect'
+export * from './web3modal/pure/AccountIndexSelect'
 
 // Utils
 export * from './api/utils/connection'
@@ -24,3 +24,4 @@ export * from './api/utils/connection'
 // TODO: this export is discussable, however it's already used outside
 export * from './api/state'
 export * from './api/state'
+export { useOpenWalletModal } from './web3modal/hooks/useOpenWalletModal'
