+++ libs/wallet/src/web3modal/hooks/useIsMetaMask.ts
