+++ libs/wallet/src/web3modal/hooks/useIsSmartContractWallet.ts
