+++ libs/wallet/src/web3modal/hooks/useIsWalletConnect.ts
