+++ libs/wallet/src/web3modal/hooks/useSafeAppsSdk.ts
