+++ libs/wallet/src/web3modal/hooks/useSafeAppsSdkInfo.ts
