+++ libs/wallet/src/web3modal/hooks/useWalletMetadata.ts
